Add controller to process the starting / sockets of terminal requests
